Simplify and consolidate to just the code we need
